move run_star_align to a dedicated tasks module
